refactor: consolidate src/data-sources with src/main.ts
