--- src/data-sources/index.ts
@@ -1,1 +1,0 @@
-export * from './data-source-registration.js';
